Fix type mismatch: path_rank INT to BIGINT
- Change path_rank return type from INT to BIGINT in both functions
- analyze_portfolio_copy_paths: path_rank INT → BIGINT
- analyze_creator_copy_paths: path_rank INT → BIGINT
- Fixes: "Returned type bigint does not match expected type integer"
- ROW_NUMBER() returns BIGINT by default in PostgreSQL

+++ supabase/migrations/20251204_fix_ambiguous_column_references.sql
@@ -9,7 +9,7 @@
 CREATE OR REPLACE FUNCTION analyze_portfolio_copy_paths()
 RETURNS TABLE(
   analysis_type TEXT,
-  path_rank INT,
+  path_rank BIGINT,
   portfolio_sequence TEXT[],
   converter_count INT,
   pct_of_converters NUMERIC,
@@ -169,7 +169,7 @@ $$;
 CREATE OR REPLACE FUNCTION analyze_creator_copy_paths()
 RETURNS TABLE(
   analysis_type TEXT,
-  path_rank INT,
+  path_rank BIGINT,
   creator_sequence TEXT[],
   converter_count INT,
   pct_of_converters NUMERIC,
